# entity-service/openapi.yaml
openapi: 3.0.1
info:
  title: Entity Service API
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /comments:
    post:
      summary: Create a comment on any supported reference entity.
      description: >-
        The reference-generic comment-create path, used for reference types that have
        no dedicated create-comment route of their own: change_request, incident,
        conversation and deployment. referenceType `case` is rejected with 400: case
        comments are owned by POST /cases/{id}/comments, which applies the
        case-specific rules.
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        "200":
          description: The created comment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCommentResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Referenced entity not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /comments/search:
    post:
      summary: Search comments by reference entity (ServiceNow data source only).
      operationId: searchComments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCommentsRequest'
      responses:
        "200":
          description: Paginated list of comments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCommentsResponse'
        "400":
          description: Bad request (missing or invalid referenceId/referenceType).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incidents/{id}:
    get:
      summary: Get an incident by ID (ServiceNow data source only).
      operationId: getIncident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Incident detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentView'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Partially update an incident (ServiceNow data source only).
      operationId: updateIncident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentRequest'
      responses:
        "200":
          description: Incident updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateIncidentResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incidents/{id}/activities/search:
    post:
      summary: >-
        Search the activity feed of an incident (comments, attachments, and optional
        field changes). Confirmed as a distinct endpoint from the case one, built on the
        same underlying mechanism.
      operationId: searchIncidentActivities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Incident UUID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchIncidentActivitiesRequest'
      responses:
        "200":
          description: Paginated activity feed for the incident.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchIncidentActivitiesResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incidents:
    post:
      summary: Create a new incident (ServiceNow data source only).
      operationId: createIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
      responses:
        "201":
          description: Incident created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIncidentResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Referenced resource not found (e.g. caller, service, or assignment group).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incidents/search:
    post:
      summary: Search incidents (ServiceNow data source only).
      operationId: searchIncidents
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchIncidentsRequest'
      responses:
        "200":
          description: Paginated list of incidents matching the filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchIncidentsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateCommentRequest:
      type: object
      required: [referenceId, referenceType, type, content]
      properties:
        referenceId:
          type: string
          format: uuid
          description: UUID of the entity the comment belongs to.
        referenceType:
          type: string
          enum: [conversation, change_request, deployment, incident]
          description: >-
            Deliberately a strict subset of ReferenceType: `case` is rejected with
            400 here, because case comments are owned by POST /cases/{id}/comments.
            Declared inline rather than referencing the shared ReferenceType, which
            must keep accepting `case` for attachments and comment search.
        type:
          type: string
          enum: [work_note, comment, activity]
          description: work_note is restricted to internal users; activity to internal and system users.
        content:
          type: string
    CreateCommentResponse:
      type: object
      properties:
        message:
          type: string
        comment:
          $ref: '#/components/schemas/CommentDetail'
...
    SearchCommentsRequest:
      type: object
      required:
        - referenceId
        - referenceType
      properties:
        referenceId:
          type: string
          format: uuid
          description: UUID of the entity to fetch comments for.
        referenceType:
          type: string
          enum: [case, conversation, change_request, deployment]
          description: Type of the referenced entity.
        pagination:
          $ref: '#/components/schemas/Pagination'
    SearchCommentsResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
          description: List of comments.
        total:
          type: integer
          description: Total number of comments matching the query.
        limit:
          type: integer
          description: Maximum number of comments returned in this page.
        offset:
          type: integer
          description: Number of comments skipped before this page.
        hasMore:
          type: boolean
          description: Whether additional pages are available.
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...
    UpdateIncidentRequest:
      type: object
      description: All fields are optional, but at least one must be provided.
      minProperties: 1
      properties:
        subject:
          type: string
        priority:
          type: string
          enum: [CRITICAL, HIGH, MODERATE, LOW, PLANNING]
        state:
          type: string
          enum: [NEW, IN_PROGRESS, ON_HOLD, RESOLVED, CLOSED, CANCELLED]
        category:
          type: string
          enum: [INQUIRY, SERVICE_INTERRUPTION, SECURITY]
        subcategory:
          type: string
          enum: [DHCP, ORACLE, CPU, KEYBOARD, DOS_DDOS, PRIVILEGE_ESCALATIONS, THREAT_INTELLIGENCE,
                 SCANS_AND_PROBES, APPLICATION_SECURITY, CONFIG_CHANGE_REQUEST, IP_ADDRESS,
                 FULL_OUTAGE, SQL_SERVER, SLOWNESS, MEMORY, MOUSE, PRIVACY, DATA_BREACH,
                 SYSTEM_COMPROMISES, DNS, OS, DISK, VPN, MALWARE, VULNERABILITY,
                 UNAUTHORIZED_ACCESS, IDENTITY_PROTECTION, PHISHING, IMPROPER_CONFIGURATION,
                 INFORMATION_REQUEST, DB2, PARTIAL_OUTAGE, EMAIL, MONITOR, WIRELESS]
        contactType:
          type: string
          enum: [SELF_SERVICE, EMAIL, WALK_IN, AZURE, EMAIL_INTERNAL, SITE_247, DIRECT,
                 PHONE, SENTINEL, VIRTUAL_AGENT, CHAT, EMAIL_EXTERNAL]
        impact:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        urgency:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        resolutionCode:
          type: string
          nullable: true
        parentId:
          type: string
          format: uuid
          nullable: true
        parentIncidentId:
          type: string
          format: uuid
          nullable: true
        assignmentGroupId:
          type: string
          format: uuid
          nullable: true
        assignedEngineerId:
          type: string
          format: uuid
          nullable: true
        serviceId:
          type: string
          format: uuid
          nullable: true
        serviceOfferingId:
          type: string
          format: uuid
          nullable: true
        configurationItemId:
          type: string
          format: uuid
          nullable: true
        changeRequestId:
          type: string
          format: uuid
          nullable: true
        problemId:
          type: string
          format: uuid
          nullable: true
        causedById:
          type: string
          format: uuid
          nullable: true
        resolvedById:
          type: string
          format: uuid
          nullable: true
        resolutionNotes:
          type: string
          nullable: true
        incidentReport:
          type: string
          nullable: true
        additionalComments:
          type: string
          nullable: true
        workNotes:
          type: string
          nullable: true
        watchList:
          type: array
          items:
            type: string
...